requirements.txt
pandas
numpy
scipy
fastapi
uvicorn
streamlit
plotly
requests
pyyaml
scikit-learn
pyarrow
